Create selenium_searching_script.py
Hello, just made a small script using selenium module it can be used inside some projects to automate stuff if anyone needed. To use this you just need to install selenium library in python and download chromewebdriver to run this program. Thanks! It feels great to contribute to open source.