Program to carry out infix to postfix conversion.
Implementation of a program to carry out infix to postfix
conversion.
Theory:
The Polish Mathematician Han Lukasiewicz suggested a notation called Polish
notation, which gives two alternatives to represent an arithmetic expression,
namely the postfix and prefix notations. In postfix notation, the operator is written
after the two operands. The advantages of the postfix notation (also called reverse
polish notation) are:
1. The need for parenthesis as in an infix expression is overcome in postfix and
prefix notations.
2. The priority of operators is no longer relevant.
3. The order of evaluation depends on the position of the operator but not on
priority and associativity.
Algorithm to convert infix expression to postfix:
1. Scan the infix expression from left to right.
2. If the scanned character is an operand, output it.
3. Else,
3.1 If the ICP>ISP or the stack is empty, push it.
3.2 Else, pop the operator from the stack until ICP<=ISP. Push the scanned
operator to the stack.
4. If the scanned character is an “(“push it onto the stack.
5. If the scanned character is an “)”, pop and output from the stack until an “(” is
encountered.
6. Repeat steps 2-6 until the infix expression is totally scanned.
7. Pop and output from the stack until it is not empty.